Refactor endpoints model for API Particulier
