+++ app/models/api_entreprise/endpoint.rb
@@ -18,6 +18,14 @@ class APIEntreprise::Endpoint < AbstractEndpoint
     open_api_definition['x-maintenances']
   end
 
+  def root_links
+    @root_links ||= extract_root_properties_from_schema('links')
+  end
+
+  def root_meta
+    @root_meta ||= extract_root_properties_from_schema('meta')
+  end
+
   def custom_provider_errors
     @custom_provider_errors ||= error_examples('502').reject do |error_payload|
       %w[
@@ -67,5 +75,33 @@ class APIEntreprise::Endpoint < AbstractEndpoint
 
   def implemented?
     !dummy?
+  end
+
+  def extract_data_from_schema
+    properties_path = %w[properties data properties]
+    properties_path.insert(2, 'items') if collection?
+    properties_path.insert(-1, 'data') if collection?
+    properties_path.insert(-1, 'properties') if collection?
+
+    response_schema.dig(*properties_path) || {}
+  end
+
+  def extract_properties_from_schema(name)
+    properties_path = ['properties', 'data', 'properties', name]
+    properties_path.insert(2, 'items') if collection?
+
+    response_schema.dig(*properties_path).try(:[], 'properties') || {}
+  end
+
+  def extract_root_properties_from_schema(name)
+    response_schema.dig('properties', name).try(:[], 'properties') || {}
+  end
+
+  def response_schema
+    ok_response = open_api_definition['responses']['200']
+
+    return if ok_response.blank?
+
+    ok_response['content']['application/json']['schema']
   end
 end
