+++ app/models/api_particulier/endpoint.rb
@@ -10,4 +10,9 @@ class APIParticulier::Endpoint < AbstractEndpoint
   def collection?
     false
   end
+
+  def extract_data_from_schema
+    data_attributes_to_dig = %w[responses 200 content application/json schema properties]
+    open_api_definition.dig(*data_attributes_to_dig)
+  end
 end
